Moved the AccessDeniedAuthorizeAttribute to the Filters namespace.
